+++ ChangeLog.md
@@ -17,6 +17,13 @@ See docs/process.md for how version tagging works.
 
 Current Trunk
 -------------
+- Only strip the LLVM producer's section in release builds. In `-O0` builds, we
+  try to leave the wasm from LLVM unmodified as much as possible, so if it
+  emitted the producers section, it will be there. Normally that only matters
+  in release builds, which is not changing here. If you want to not have a
+  producer's section in debug builds, you can remove it a tool like
+  `wasm-opt --strip-producers` (which is what Emscripten still does in release
+  builds, as always) or use `llvm-objcopy`.
 - Do not remove `__original_main` using `--inline-main`. We used to do this
   so that it didn't show up in stack traces (which could be confusing because
   it is added by the linker - it's not in the source code). But this has had